modified game mechanics apis; added dice to player_turn table - added dice to player_turn table - added index on status in game table - added game rankings controller and route - added LIMIT to SELECT queries - added checks for game start / exist / finish - added checks for matching request header and db values

+++ backend/controllers/gameController.js
@@ -14,7 +14,7 @@ const create_game = async (req, res) => {
   try {
     await client.query("BEGIN");
     const availableGames = await client.query(
-      "SELECT * FROM game WHERE status='WAITING_FOR_PLAYERS' ORDER BY created_on DESC"
+      "SELECT * FROM game WHERE status='WAITING_FOR_PLAYERS' ORDER BY created_on DESC LIMIT 1"
     );
 
     let game;
@@ -78,7 +78,7 @@ const create_game = async (req, res) => {
       ]);
     }
 
-    const gameData = await client.query("SELECT * FROM game WHERE id = $1", [
+    const gameData = await client.query("SELECT * FROM game WHERE id = $1 LIMIT 16", [
       gameId,
     ]);
 
@@ -99,25 +99,62 @@ const get_game_state = async (req, res) => {
   try {
     await client.query("BEGIN");
     const gameId = parseInt(req.headers.game_id);
-    const coinStates = await client.query(
-      "SELECT coin_state.id, coin_state.color, coin_state.position, coin_state.player_id FROM coin_state JOIN player ON coin_state.player_id = player.id WHERE player.game_id = $1",
+
+    // checking if game has not started
+    const gameStatus = await client.query(
+      "SELECT status FROM game WHERE id = $1 LIMIT 1",
       [gameId]
     );
 
-    // checking if game has ended
+    // checking if requested game-id exists
+    if (gameStatus.rowCount === 0) {
+      await client.query("ROLLBACK");
+      const error = "requested game-id does not exist.";
+      console.error(error);
+      res.status(500).json({ error: error });
+      return;
+    }
+
+    if (gameStatus.rows[0].status === "WAITING_FOR_PLAYERS") {
+      await client.query("ROLLBACK");
+      const error = "requested game-id has not started.";
+      console.error(error);
+      res.status(500).json({ error: error });
+      return;
+    }
+
+    // checking if game has finished
     const finishPlayerCount = await client.query(
-      "SELECT COUNT(id) FROM player WHERE game_id = $1 AND status = 'FINISHED'",
+      "SELECT COUNT(id) FROM player WHERE game_id = $1 AND status = 'FINISHED' LIMIT 4",
       [gameId]
     );
     const gameHasEnded = finishPlayerCount.rows[0].count === 3;
 
-    // if game has ended: dropping game from player-turn table
+    // if game has finished
     if (gameHasEnded) {
+      // dropping game from player-turn table
       await client.query("DELETE from player_turn where game_id = $1", [
         gameId,
       ]);
+
+      // updating game status
+      await client.query(
+        "UPDATE game SET status = 'FINISHED' where game_id = $1",
+        [gameId]
+      );
+
+      // updating finish-timestamp for the last player
+      await client.query(
+        "UPDATE player SET status = 'FINISHED' where game_id = $1 and status <> 'FINISHED'",
+        [gameId]
+      );
+      res.status(200).json({ board_state: null, game_has_ended: gameHasEnded });
     }
 
+    const coinStates = await client.query(
+      "SELECT coin_state.id, coin_state.color, coin_state.position, coin_state.player_id FROM coin_state JOIN player ON coin_state.player_id = player.id WHERE player.game_id = $1 LIMIT 16",
+      [gameId]
+    );
     const coinStateData = coinStates.rows;
 
     await client.query("COMMIT");
@@ -139,78 +176,127 @@ const roll_dice = async (req, res) => {
   try {
     await client.query("BEGIN");
     const gameId = req.headers.game_id;
-    
-    // ? get player-id from player-turn table directly
-    const playerId = req.headers.player_id;
 
-    // checking if game has ended
-    const playerTurn = await client.query(
-      "SELECT COUNT(id) FROM player_turn WHERE game_id = $1",
+    // checking if game has not started
+    const gameStatus = await client.query(
+      "SELECT status FROM game WHERE id = $1 LIMIT 1",
       [gameId]
     );
 
-    const gameHasEnded = playerTurn.rows[0].count === 0;
+    // checking if requested game-id exists
+    if (gameStatus.rowCount === 0) {
+      await client.query("ROLLBACK");
+      const error = "requested game-id does not exist.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
 
-    // if game has not ended
-    if (!gameHasEnded) {
-      // generating random number from 1 to 6
-      const diceValue = Math.floor(Math.random() * 6) + 1;
+    if (gameStatus.rows[0].status === "WAITING_FOR_PLAYERS") {
+      await client.query("ROLLBACK");
+      const error = "requested game-id has not started.";
+      console.error(error);
+      res.status(500).json({ error: error });
+      return;
+    }
 
-      // checking if any moves possible (coin-position + dice-value <= 56)
-      const value = 56 - diceValue;
+    const gameHasEnded = gameStatus.rows[0].status === "FINISHED";
 
-      let moveableCoins;
-      if (diceValue === 6) {
-        moveableCoins = await client.query(
-          "SELECT COUNT(id) FROM coin_state WHERE player_id = $1 AND position <= $2",
-          [playerId, value]
-        );
-      } else {
-        moveableCoins = await client.query(
-          "SELECT COUNT(id) FROM coin_state WHERE player_id = $1 AND position <= $2 AND position <> -1",
-          [playerId, value]
-        );
-      }
-      const countMoveableCoins = parseInt(moveableCoins.rows[0].count);
-      const cannotMove = countMoveableCoins === 0;
-
-      // if no move possible; update player turn
-      if (cannotMove) {
-        // getting unfinished players (as well as current player in order)
-        const unfinishedPlayers = await client.query(
-          "SELECT * FROM player WHERE (game_id = $1 AND status = 'IN_GAME') OR id = $2 ORDER BY id",
-          [gameId, playerId]
-        );
-        const unfinishedPlayersOrder = unfinishedPlayers.rows.map(
-          (row) => row.id
-        );
-        let nextTurnPlayerId = null;
-        const numPlayers = unfinishedPlayersOrder.length;
-        if (numPlayers !== 1) {
-          const nextPlayerIndex =
-            (unfinishedPlayersOrder.indexOf(playerId) + 1) % numPlayers;
-          nextTurnPlayerId = unfinishedPlayersOrder[nextPlayerIndex];
-        }
-
-        if (nextTurnPlayerId)
-          await client.query(
-            "UPDATE player_turn SET player_id = $1 WHERE game_id = $2",
-            [nextTurnPlayerId, gameId]
-          );
-      }
-
-      await client.query("COMMIT");
-      res.status(200).json({
-        game_has_ended: false,
-        dice_value: diceValue,
-        moves_possible: !cannotMove,
-      });
-    } else {
+    // checking if game has finished
+    if (gameHasEnded) {
       await client.query("COMMIT");
       res.status(200).json({
         game_has_ended: true,
       });
+      return;
     }
+
+    // get player-id and dice-value from player-turn table directly
+    const playerTurn = await client.query(
+      "SELECT player_id, dice FROM player_turn WHERE game_id = $1 LIMIT 1",
+      [gameId]
+    );
+
+    // checking if previous turn player had moved their coin
+    if (playerTurn.rows[0].dice !== 0) {
+      await client.query("ROLLBACK");
+      const error = "Previous player did not move their coin.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
+
+    // checking if player-id in request header matches player-id in player-turn
+    const playerId = playerTurn.rows[0].player_id;
+    const reqPlayerId = req.headers.player_id;
+
+    if (playerId !== reqPlayerId) {
+      await client.query("ROLLBACK");
+      const error =
+        "player-id in player-turn table does not match the player-id in request header.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
+
+    // generating random number from 1 to 6
+    const diceValue = Math.floor(Math.random() * 6) + 1;
+
+    // checking if any moves possible (coin-position + dice-value <= 56)
+    const value = 56 - diceValue;
+
+    let moveableCoins;
+    if (diceValue === 6) {
+      moveableCoins = await client.query(
+        "SELECT COUNT(id) FROM coin_state WHERE player_id = $1 AND position <= $2 AND position <> 56 LIMIT 4",
+        [playerId, value]
+      );
+    } else {
+      moveableCoins = await client.query(
+        "SELECT COUNT(id) FROM coin_state WHERE player_id = $1 AND position <= $2 AND position <> -1 AND position <> 56 LIMIT 4",
+        [playerId, value]
+      );
+    }
+    const countMoveableCoins = parseInt(moveableCoins.rows[0].count);
+    const cannotMove = countMoveableCoins === 0;
+
+    // if no move possible; update player turn
+    if (cannotMove) {
+      // getting unfinished players (as well as current player in order)
+      const unfinishedPlayers = await client.query(
+        "SELECT id FROM player WHERE game_id = $1 AND status = 'IN_GAME' ORDER BY id LIMIT 4",
+        [gameId]
+      );
+      const unfinishedPlayersOrder = unfinishedPlayers.rows.map(
+        (row) => row.id
+      );
+      let nextTurnPlayerId = null;
+      const numPlayers = unfinishedPlayersOrder.length;
+      if (numPlayers !== 1) {
+        const nextPlayerIndex =
+          (unfinishedPlayersOrder.indexOf(playerId) + 1) % numPlayers;
+        nextTurnPlayerId = unfinishedPlayersOrder[nextPlayerIndex];
+      }
+
+      if (nextTurnPlayerId)
+        await client.query(
+          "UPDATE player_turn SET player_id = $1, dice = 0 WHERE game_id = $2",
+          [nextTurnPlayerId, gameId]
+        );
+    } else {
+      // update dice value in player-turn table
+      await client.query(
+        "UPDATE player_turn SET player_id = $1, dice = $2 WHERE game_id = $3",
+        [playerId, diceValue, gameId]
+      );
+    }
+
+    await client.query("COMMIT");
+    res.status(200).json({
+      game_has_ended: false,
+      dice_value: diceValue,
+      moves_possible: !cannotMove,
+    });
   } catch (error) {
     await client.query("ROLLBACK");
     console.error("Error in rolling the dice:", error);
@@ -227,15 +313,65 @@ const move_coin = async (req, res) => {
     await client.query("BEGIN");
 
     const coinId = req.headers.coin_id;
-    const coinColor = req.headers.coin_color;
-    const playerId = req.headers.player_id;
     const gameId = req.headers.game_id;
-    const diceValue = req.headers.dice_value;
+
+    const coinData = await client.query(
+      "SELECT color, player_id FROM coin_state WHERE id = $1 LIMIT 1",
+      [coinId]
+    );
+    const coinColor = coinData.rows[0].color;
+    const playerId = parseInt(coinData.rows[0].player_id);
+
+    const playerTurn = await client.query(
+      "SELECT dice, player_id from player_turn where game_id = $1 LIMIT 1",
+      [gameId]
+    );
+
+    // checking if game-id exists or not
+    if (playerTurn.rowCount === 0) {
+      await client.query("ROLLBACK");
+      const error = "requested game-id does not exist.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
+
+    const diceValue = playerTurn.rows[0].dice;
+
+    // check if the current player rolled the dice
+    if (diceValue === 0) {
+      await client.query("ROLLBACK");
+      const error = "previous player did not roll the dice.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
+
+    // checking if dice-value in request header matches dice-value in player-turn
+    const reqDiceValue = parseInt(req.headers.dice_value);
+    if (diceValue !== reqDiceValue) {
+      await client.query("ROLLBACK");
+      const error =
+        "dice-value in player-turn table does not match the dice-value in request header.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
+
+    // checking if requested coin id belongs to player in player-turn
+    if (playerId !== parseInt(playerTurn.rows[0].player_id)) {
+      await client.query("ROLLBACK");
+      const error =
+        "coin-id in request header does not belong to player-id in player-turn table.";
+      console.error(error);
+      res.status(401).json({ error: error });
+      return;
+    }
 
     // checking if the requested coin can move or not
 
     let coinPosition = await client.query(
-      "SELECT position FROM coin_state WHERE id = $1",
+      "SELECT position FROM coin_state WHERE id = $1 LIMIT 1",
       [coinId]
     );
     coinPosition = coinPosition.rows[0].position;
@@ -244,17 +380,17 @@ const move_coin = async (req, res) => {
       (coinPosition !== -1 && coinPosition + diceValue <= 56) |
       (coinPosition === -1 && diceValue === 6)
     ) {
-      coinPosition += diceValue;
+      coinPosition = coinPosition === -1 ? 0 : coinPosition + diceValue;
 
       // updating position in coin state
-      await client.query(
-        "UPDATE coin_state SET position = $1 WHERE coin_id = $2",
-        [coinPosition, coinId]
-      );
+      await client.query("UPDATE coin_state SET position = $1 WHERE id = $2", [
+        coinPosition,
+        coinId,
+      ]);
 
       // getting unfinished players (as well as current player in order)
       const unfinishedPlayers = await client.query(
-        "SELECT * FROM player WHERE (game_id = $1 AND status = 'IN_GAME') OR player_id = $2 ORDER BY id",
+        "SELECT id FROM player WHERE game_id = $1 AND status = 'IN_GAME' ORDER BY id LIMIT 4",
         [gameId, playerId]
       );
       const unfinishedPlayersOrder = unfinishedPlayers.rows.map(
@@ -286,12 +422,16 @@ const move_coin = async (req, res) => {
           // update next player's turn
           if (nextTurnPlayerId !== null) {
             await client.query(
-              "UPDATE player_turn SET player_id = $1 where game_id = $1",
+              "UPDATE player_turn SET player_id = $1, dice = 0 where game_id = $1",
               [nextTurnPlayerId, gameId]
             );
           }
         }
-        // otherwise no change in player turn
+        // otherwise no change in player turn but update dice-value in player-turn table
+        await client.query(
+          "UPDATE player_turn SET dice = 0 where game_id = $1",
+          [gameId]
+        );
 
         // coin does not reach home
       } else {
@@ -301,7 +441,7 @@ const move_coin = async (req, res) => {
           [absoluteCoinPosition, coinId]
         );
         const hasCut = cut.rowCount !== 0;
-        // if new coin position is among the safe positions
+        // if new coin position is not among the safe positions
         if (!safePositions.has(coinPosition) && hasCut) {
           const cutCoinId = cut.rows[0].id;
           // set the cut coin position to -1
@@ -309,31 +449,35 @@ const move_coin = async (req, res) => {
             "UPDATE coin_state SET position = -1 WHERE id = $1",
             [cutCoinId]
           );
-          // otherwise no change in turn
         } else if (diceValue !== 6) {
           // next player's turn update
           if (nextTurnPlayerId !== null) {
             await client.query(
-              "UPDATE player_turn SET player_id = $1 where game_id = $1",
+              "UPDATE player_turn SET player_id = $1, dice = 0 where game_id = $2",
               [nextTurnPlayerId, gameId]
             );
           }
         }
-      }
 
-      // ? return board state
-      await client.query("COMMIT");
-      res.status(200).json({ can_move: true });
-    } else {
-      // update turn to next player
-      if (nextTurnPlayerId !== null) {
+        // update dice value to 0 in player-turn table (indicating that player has moved)
         await client.query(
-          "UPDATE player_turn SET player_id = $1 where game_id = $1",
-          [nextTurnPlayerId, gameId]
+          "UPDATE player_turn SET dice = 0 where game_id = $1",
+          [gameId]
         );
       }
+
+      // getting board state
+      const coinStates = await client.query(
+        "SELECT coin_state.id, coin_state.color, coin_state.position, coin_state.player_id FROM coin_state JOIN player ON coin_state.player_id = player.id WHERE player.game_id = $1",
+        [gameId]
+      );
+      const coinStateData = coinStates.rows;
+
       await client.query("COMMIT");
-      res.status(200).json({ can_move: false });
+      res.status(200).json({ can_move: true, board_state: coinStateData });
+    } else {
+      await client.query("COMMIT");
+      res.status(200).json({ can_move: false, board_state: null });
     }
   } catch (error) {
     await client.query("ROLLBACK");
@@ -344,4 +488,51 @@ const move_coin = async (req, res) => {
   }
 };
 
-module.exports = { create_game, get_game_state, roll_dice, move_coin };
+const get_game_rankings = async (req, res) => {
+  const client = await pool.connect();
+
+  try {
+    await client.query("BEGIN");
+    const gameId = parseInt(req.headers.game_id);
+
+    // checking if game has finished
+    const gameStatus = await client.query(
+      "SELECT status from game where id = $1 LIMIT 1",
+      [gameId]
+    );
+
+    // checking if requested game-id exists or not
+    if (gameStatus.rowCount === 0) {
+      await client.query("ROLLBACK");
+      const error = "requested game-id does not exist.";
+      console.error(error);
+      res.status(500).json({ error: error });
+      return;
+    }
+
+    if (gameStatus.rows[0].status === "FINISHED") {
+      const rankings = await client.query(
+        "SELECT id FROM player where game_id = $1 ORDER_BY finished_ts LIMIT 4",
+        [gameId]
+      );
+      res.status(200).json({ rankings: rankings.rows });
+    }
+
+    await client.query("COMMIT");
+    res.status(200).json({ rankings: null });
+  } catch (error) {
+    await client.query("ROLLBACK");
+    console.error("Error in generating the game rankings: ", error);
+    res.status(500).json({ error: error });
+  } finally {
+    client.release();
+  }
+};
+
+module.exports = {
+  create_game,
+  get_game_state,
+  roll_dice,
+  move_coin,
+  get_game_rankings,
+};
